+++ src/main/java/org/omarsalem/gameel/birds/services/SimilarityCalculator.java
@@ -3,5 +3,5 @@ package org.omarsalem.gameel.birds.services;
 import java.util.Map;
 
 public interface SimilarityCalculator {
-    double calculateScore(Map<Integer, Integer> userPreferences, Map<Integer, Integer> otherUserPreferences);
+    double calculateScore(Map<Integer, Double> userPreferences, Map<Integer, Double> otherUserPreferences);
 }
